[FIXED JENKINS-29134]
Added support for automatically overriding the values of "BUILDS_TODAY",
"BUILDS_THIS_MONTH", "BUILDS_THIS_YEAR" and "BUILDS_ALL_TIME" with
values taken from environment-variables.

Instead of just providing a simple number in the form-fields of the
job's plugin-configuration which overrides the value for the next build,
one can now provide an environment-variable whose value will be
extracted and used instead during the next builds.
If it is not set or its value is not convertible to an integer, the
value of the previous build will be taken instead and increased by one.

Additonally, negative numbers are allowed now. (However, even they will
always only be increased and never decreased!)

Signed-off-by: Deniz Bahadir <dbahadir@benocs.com>